feat: shared loadCommands() loader for command discovery
Extract the slash-command discovery duplicated between src/index.js
and src/deploy-commands.js into src/loaders/commands.js exporting
async loadCommands(). The loader walks src/commands/<group>/*.js,
validates the data/execute exports with the same warning as before,
and returns the validated modules; index.js fills client.commands
from it while deploy-commands.js maps to .data.toJSON().

Closes #17
